refactor: CropViewModel의 메서드를 UseCase, Repository로 분리
